Uses only Armor & Ammo
Only modification is removal of random objects that I've never seen used on any of the FiveM servers, and adds more configurations such as whether your character has to run an animation before the armor is equipped, or if it comes on instantly. This would have been a cool feature in the original version as sometimes the animations can bug out and the character can freeze. Another Configuration I have added is if the player should actually wear a component which is by default Black Bulletproof Armor.

I have removed the trigger events and the insertion of everything that isn't ammoclip/bulletproof-vest from the client & server sided files as well as the SQL file so that less Events are being called upon.

I removed firstaidkit aswell due to it performing the same function as esx_ambulancejob's "medikit".

The original version of this upload can be found at: https://github.com/mirrox1337/esx_extraitems

Credit to @mirrox1337, thanks!